Finished refactoring and correcting some tests, mostly negative tests
