+++ src/test/java/com/danielvm/destiny2bot/integration/BaseIntegrationTest.java
@@ -18,12 +18,9 @@ import org.skyscreamer.jsonassert.JSONAssert;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.boot.web.reactive.context.ReactiveWebApplicationContext;
 import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec;
 import org.springframework.web.reactive.function.BodyInserters;
@@ -42,12 +39,6 @@ public abstract class BaseIntegrationTest {
   private static final GenericContainer<?> REDIS_CONTAINER = new GenericContainer<>(
       "redis:5.0.3-alpine").withExposedPorts(6379);
 
-  private static final String MALICIOUS_PRIVATE_KEY = "CE4517095255B0C92D586AF9EEC27B998D68775363F9FE74341483FB3A657CEC";
-  private static final String VALID_PRIVATE_KEY = "F0EA3A0516695324C03ED552CD5A08A58CA1248172E8816C3BF235E52E75A7BF";
-
-  @LocalServerPort
-  protected int localServerPort;
-
   @Autowired
   ReactiveWebApplicationContext reactiveWebApplicationContext;
 
@@ -64,18 +55,6 @@ public abstract class BaseIntegrationTest {
   BungieConfiguration bungieConfiguration;
 
   /**
-   * Starts up some dynamic properties that change because of TestContainers usage, as well as the
-   * test environment having a random port
-   *
-   * @param registry {@link DynamicPropertyRegistry}
-   */
-  @DynamicPropertySource
-  public static void setup(DynamicPropertyRegistry registry) {
-    registry.add("spring.data.redis.port", REDIS_CONTAINER::getFirstMappedPort);
-    registry.add("spring.data.redis.host", REDIS_CONTAINER::getHost);
-  }
-
-  /**
    * Asserts a Json input is valid/invalid regardless of formatting, spacing, new-lines, etc.
    *
    * @param target   The target json to be compared
@@ -87,78 +66,5 @@ public abstract class BaseIntegrationTest {
     } catch (JSONException e) {
       throw new RuntimeException(e);
     }
-  }
-
-  /**
-   * Sends a partial WebTestClient request to the specified route with a valid signature
-   *
-   * @param endpoint    The /endpoint to call
-   * @param interaction The interaction body that the user will send along this request
-   * @return {@link WebTestClient.RequestHeadersSpec}
-   * @throws DecoderException        If something goes wrong with the Crypto library
-   * @throws JsonProcessingException If Jackson is unable to parse the body to Json for signing
-   */
-  public ResponseSpec sendValidSignatureRequest(String endpoint, Interaction interaction)
-      throws DecoderException, JsonProcessingException {
-    String timestamp = String.valueOf(Instant.now().getEpochSecond());
-    String signature = createValidSignature(interaction, timestamp);
-    return this.webTestClient.post().uri(endpoint)
-        .accept(MediaType.APPLICATION_JSON, MediaType.MULTIPART_FORM_DATA)
-        .contentType(MediaType.APPLICATION_JSON)
-        .header("X-Signature-Ed25519", signature)
-        .header("X-Signature-Timestamp", timestamp)
-        .body(BodyInserters.fromValue(interaction))
-        .exchange();
-  }
-
-  /**
-   * Sends a WebTestClient request with the appropriate headers but with an invalid signature
-   *
-   * @param endpoint    The /endpoint to call
-   * @param interaction The interaction body that the user will send along this request
-   * @return {@link WebTestClient.ResponseSpec}
-   * @throws DecoderException        If something goes wrong with the Crypto library
-   * @throws JsonProcessingException If Jackson is unable to parse the body to Json for signing
-   */
-  public ResponseSpec sendInvalidSignatureRequest(String endpoint, Interaction interaction)
-      throws DecoderException, JsonProcessingException {
-    String timestamp = String.valueOf(Instant.now().getEpochSecond());
-    String signature = createInvalidSignature(interaction, timestamp);
-    return this.webTestClient.post().uri(endpoint)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .header("X-Signature-Ed25519", signature)
-        .header("X-Signature-Timestamp", timestamp)
-        .body(BodyInserters.fromValue(interaction))
-        .exchange();
-  }
-
-  private String createValidSignature(Interaction body, String timestamp)
-      throws JsonProcessingException, DecoderException {
-    KeyPair signingKeys = Crypto.seedSigningKeyPair(Hex.decodeHex(VALID_PRIVATE_KEY.toCharArray()));
-
-    discordConfiguration.setBotPublicKey(
-        Hex.encodeHexString(
-            signingKeys.getPublic().getEncoded())); // change the public key in the config class
-
-    var signatureBytes = Crypto.sign(signingKeys.getPrivate(),
-        (timestamp + objectMapper.writeValueAsString(body)).getBytes(StandardCharsets.UTF_8));
-    return Hex.encodeHexString(signatureBytes);
-  }
-
-  private String createInvalidSignature(Interaction body, String timestamp)
-      throws JsonProcessingException, DecoderException {
-    KeyPair invalidSigningKeyPair = Crypto.seedSigningKeyPair(
-        Hex.decodeHex(MALICIOUS_PRIVATE_KEY.toCharArray()));
-
-    PublicKey validPublicKey = Crypto.seedSigningKeyPair(
-        Hex.decodeHex(VALID_PRIVATE_KEY.toCharArray())).getPublic();
-
-    discordConfiguration.setBotPublicKey(
-        Hex.encodeHexString(validPublicKey.getEncoded()));
-
-    var signatureBytes = Crypto.sign(invalidSigningKeyPair.getPrivate(),
-        (timestamp + objectMapper.writeValueAsString(body)).getBytes(StandardCharsets.UTF_8));
-    return Hex.encodeHexString(signatureBytes);
   }
 }
