+++ src/test/java/com/danielvm/destiny2bot/integration/InteractionControllerTest.java
@@ -28,21 +28,30 @@ import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.security.KeyPair;
+import java.security.PublicKey;
+import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import org.apache.commons.codec.DecoderException;
+import org.apache.commons.codec.binary.Hex;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec;
+import org.springframework.web.reactive.function.BodyInserters;
+import software.pando.crypto.nacl.Crypto;
 
 public class InteractionControllerTest extends BaseIntegrationTest {
+
+  private static final String MALICIOUS_PRIVATE_KEY = "CE4517095255B0C92D586AF9EEC27B998D68775363F9FE74341483FB3A657CEC";
+  private static final String VALID_PRIVATE_KEY = "F0EA3A0516695324C03ED552CD5A08A58CA1248172E8816C3BF235E52E75A7BF";
 
   // Static mapper to be used on the @BeforeAll static method
   private static final ObjectMapper OBJECT_MAPPER = new JsonMapper.Builder(new JsonMapper())
@@ -85,6 +94,78 @@ public class InteractionControllerTest extends BaseIntegrationTest {
             entry.getValue().setEndDate(MessageUtils.NEXT_TUESDAY);
           }
         });
+  }
+
+  /**
+   * Sends a partial WebTestClient request to the specified route with a valid signature
+   *
+   * @param endpoint    The /endpoint to call
+   * @param interaction The interaction body that the user will send along this request
+   * @return {@link WebTestClient.RequestHeadersSpec}
+   * @throws DecoderException        If something goes wrong with the Crypto library
+   * @throws JsonProcessingException If Jackson is unable to parse the body to Json for signing
+   */
+  public ResponseSpec sendValidSignatureRequest(String endpoint, Interaction interaction)
+      throws DecoderException, JsonProcessingException {
+    String timestamp = String.valueOf(Instant.now().getEpochSecond());
+    String signature = createValidSignature(interaction, timestamp);
+    return this.webTestClient.post().uri(endpoint)
+        .accept(MediaType.APPLICATION_JSON, MediaType.MULTIPART_FORM_DATA)
+        .contentType(MediaType.APPLICATION_JSON)
+        .header("X-Signature-Ed25519", signature)
+        .header("X-Signature-Timestamp", timestamp)
+        .body(BodyInserters.fromValue(interaction))
+        .exchange();
+  }
+
+  /**
+   * Sends a WebTestClient request with the appropriate headers but with an invalid signature
+   *
+   * @param endpoint    The /endpoint to call
+   * @param interaction The interaction body that the user will send along this request
+   * @return {@link WebTestClient.ResponseSpec}
+   * @throws DecoderException        If something goes wrong with the Crypto library
+   * @throws JsonProcessingException If Jackson is unable to parse the body to Json for signing
+   */
+  public ResponseSpec sendInvalidSignatureRequest(String endpoint, Interaction interaction)
+      throws DecoderException, JsonProcessingException {
+    String timestamp = String.valueOf(Instant.now().getEpochSecond());
+    String signature = createInvalidSignature(interaction, timestamp);
+    return this.webTestClient.post().uri(endpoint)
+        .accept(MediaType.APPLICATION_JSON)
+        .contentType(MediaType.APPLICATION_JSON)
+        .header("X-Signature-Ed25519", signature)
+        .header("X-Signature-Timestamp", timestamp)
+        .body(BodyInserters.fromValue(interaction))
+        .exchange();
+  }
+
+  private String createValidSignature(Interaction body, String timestamp)
+      throws JsonProcessingException, DecoderException {
+    KeyPair signingKeys = Crypto.seedSigningKeyPair(Hex.decodeHex(VALID_PRIVATE_KEY.toCharArray()));
+    discordConfiguration.setBotPublicKey(
+        Hex.encodeHexString(
+            signingKeys.getPublic().getEncoded())); // change the public key in the config class
+
+    var signatureBytes = Crypto.sign(signingKeys.getPrivate(),
+        (timestamp + objectMapper.writeValueAsString(body)).getBytes(StandardCharsets.UTF_8));
+    return Hex.encodeHexString(signatureBytes);
+  }
+
+  private String createInvalidSignature(Interaction body, String timestamp)
+      throws JsonProcessingException, DecoderException {
+    KeyPair invalidSigningKeyPair = Crypto.seedSigningKeyPair(
+        Hex.decodeHex(MALICIOUS_PRIVATE_KEY.toCharArray()));
+
+    PublicKey validPublicKey = Crypto.seedSigningKeyPair(
+        Hex.decodeHex(VALID_PRIVATE_KEY.toCharArray())).getPublic();
+
+    discordConfiguration.setBotPublicKey(
+        Hex.encodeHexString(validPublicKey.getEncoded()));
+
+    var signatureBytes = Crypto.sign(invalidSigningKeyPair.getPrivate(),
+        (timestamp + objectMapper.writeValueAsString(body)).getBytes(StandardCharsets.UTF_8));
+    return Hex.encodeHexString(signatureBytes);
   }
 
   @Test
@@ -240,29 +321,19 @@ public class InteractionControllerTest extends BaseIntegrationTest {
     stubFor(get(urlPathEqualTo("/bungie/Destiny2/Milestones/"))
         .withHeader("x-api-key", equalTo(bungieConfiguration.getKey()))
         .willReturn(aResponse()
-            .withStatus(400)
+            .withStatus(200)
             .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            .withBodyFile("bungie/missing-api-key.json")));
+            .withBodyFile("bungie/no-weekly-data.json")));
 
     // when: the request is sent
     var response = sendValidSignatureRequest("/interactions", body);
 
     // then: the response JSON is correct
-    String errorJson;
-    try {
-      errorJson = objectMapper.writeValueAsString(
-          objectMapper.readValue(
-              new ClassPathResource("__files/bungie/missing-api-key.json").getInputStream(),
-              Object.class));
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-
     response.expectStatus()
-        .isBadRequest()
+        .is5xxServerError()
         .expectBody()
-        .jsonPath("$.detail").value(json -> assertJsonLenient(errorJson, json))
-        .jsonPath("$.status").isEqualTo(HttpStatus.BAD_REQUEST.value())
+        .jsonPath("$.status").isEqualTo(500)
+        .jsonPath("$.detail").isEqualTo("No available milestone data was available for processing")
         .returnResult();
   }
 
@@ -281,11 +352,24 @@ public class InteractionControllerTest extends BaseIntegrationTest {
         .data(data).build();
 
     // when: the request is sent
-    ResponseSpec response = sendInvalidSignatureRequest("/interactions", body);
+    String timestamp = String.valueOf(Instant.now().getEpochSecond());
+    String signature = createInvalidSignature(body, timestamp);
+    var response = webTestClient.post().uri("/interactions")
+        .accept(MediaType.APPLICATION_JSON)
+        .contentType(MediaType.APPLICATION_JSON)
+        .header("X-Signature-Ed25519", signature)
+        .header("X-Signature-Timestamp", timestamp)
+        .body(BodyInserters.fromValue(body))
+        .exchange();
 
     // then: the response JSON has the correct error message
     response.expectStatus()
-        .isBadRequest();
+        .isBadRequest()
+        .expectBody()
+        .jsonPath("$.status").isEqualTo(400)
+        .jsonPath("$.detail")
+        .isEqualTo("The signature passed in was invalid. Timestamp: [%s], Signature [%s]"
+            .formatted(timestamp, signature));
   }
 
   @Test
@@ -314,11 +398,24 @@ public class InteractionControllerTest extends BaseIntegrationTest {
         .build();
 
     // when: the request is sent
-    var response = sendInvalidSignatureRequest("/interactions", body);
+    String timestamp = String.valueOf(Instant.now().getEpochSecond());
+    String signature = createInvalidSignature(body, timestamp);
+    var response = webTestClient.post().uri("/interactions")
+        .accept(MediaType.APPLICATION_JSON)
+        .contentType(MediaType.APPLICATION_JSON)
+        .header("X-Signature-Ed25519", signature)
+        .header("X-Signature-Timestamp", timestamp)
+        .body(BodyInserters.fromValue(body))
+        .exchange();
 
     // then: the response JSON has the correct error message
     response.expectStatus()
-        .isBadRequest();
+        .isBadRequest()
+        .expectBody()
+        .jsonPath("$.status").isEqualTo(400)
+        .jsonPath("$.detail")
+        .isEqualTo("The signature passed in was invalid. Timestamp: [%s], Signature [%s]".formatted(
+            timestamp, signature));
   }
 
   @Test
