+++ src/test/resources/__files/bungie/milestone-response.json
@@ -1,1 +1,1 @@
-{"response":{"3021174356":{"milestoneHash":3021174356,"startDate":null,"endDate":null,"activities":null},"644555645":{"milestoneHash":644555645,"startDate":null,"endDate":null,"activities":null},"3181387331":{"milestoneHash":3181387331,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":[{"activityHash":2122313384,"challengeObjectiveHashes":[]}]},"4253138191":{"milestoneHash":4253138191,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":null},"3603098564":{"milestoneHash":3603098564,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":null},"3802603984":{"milestoneHash":3802603984,"startDate":null,"endDate":null,"activities":null},"2709491520":{"milestoneHash":2709491520,"startDate":null,"endDate":null,"activities":null},"2594202463":{"milestoneHash":2594202463,"startDate":null,"endDate":null,"activities":null},"3899487295":{"milestoneHash":3899487295,"startDate":null,"endDate":null,"activities":null},"2712317338":{"milestoneHash":2712317338,"startDate":1709658000.000000000,"endDate":1710864000.000000000,"activities":[{"activityHash":1042180643,"challengeObjectiveHashes":["2398860795"]}]},"541780856":{"milestoneHash":541780856,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":[{"activityHash":910380154,"challengeObjectiveHashes":[]}]},"526718853":{"milestoneHash":526718853,"startDate":1709658000.000000000,"endDate":1710864000.000000000,"activities":[{"activityHash":1262462921,"challengeObjectiveHashes":["3211393925"]},{"activityHash":2296818662,"challengeObjectiveHashes":["3211393925"]}]}}}
+{"response":{"3021174356":{"milestoneHash":3021174356,"startDate":null,"endDate":null,"activities":null},"644555645":{"milestoneHash":644555645,"startDate":null,"endDate":null,"activities":null},"3181387331":{"milestoneHash":3181387331,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":[{"activityHash":2122313384,"challengeObjectiveHashes":[]}]},"4253138191":{"milestoneHash":4253138191,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":null},"3603098564":{"milestoneHash":3603098564,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":null},"3802603984":{"milestoneHash":3802603984,"startDate":null,"endDate":null,"activities":null},"2709491520":{"milestoneHash":2709491520,"startDate":null,"endDate":null,"activities":null},"2594202463":{"milestoneHash":2594202463,"startDate":null,"endDate":null,"activities":null},"3899487295":{"milestoneHash":3899487295,"startDate":null,"endDate":null,"activities":null},"2712317338":{"milestoneHash":2712317338,"startDate":1710259200.000000000,"endDate":1710864000.000000000,"activities":[{"activityHash":1042180643,"challengeObjectiveHashes":["2398860795"]}]},"541780856":{"milestoneHash":541780856,"startDate":1701795600.000000000,"endDate":1702400400.000000000,"activities":[{"activityHash":910380154,"challengeObjectiveHashes":[]}]},"526718853":{"milestoneHash":526718853,"startDate":1710259200.000000000,"endDate":1710864000.000000000,"activities":[{"activityHash":1262462921,"challengeObjectiveHashes":["3211393925"]},{"activityHash":2296818662,"challengeObjectiveHashes":["3211393925"]}]}}}
