+++ src/test/resources/__files/bungie/no-weekly-data.json
@@ -1,0 +1,7 @@
+{
+  "ErrorCode": 1,
+  "ThrottleSeconds": 0,
+  "ErrorStatus": "Success",
+  "Message": "Ok",
+  "MessageData": {}
+}
